Fix logic type for evaluation CI

+++ inference/evaluationCI.py
@@ -17,7 +17,8 @@ def load_data(csv_path):
         'hate_score', 'toxicity', 'obscene', 'identity_attack', 'insult', 'threat', 'sexual_explicit',
         'sentiment_negative', 'sentiment_neutral', 'sentiment_positive',
         'emotion_anger', 'emotion_fear', 'emotion_joy', 'emotion_love', 'emotion_sadness', 'emotion_surprise'
-    ]].values
+    ]].values.astype(np.float32)
+    
     labels = df['label'].values
 
     _, val_comments, _, val_scalar, _, y_val = train_test_split(
